Meta: Rewrite GeneratePublicSuffixData in python
