+++ Meta/Lagom/Tools/CodeGenerators/CMakeLists.txt
@@ -1,3 +1,2 @@
 add_subdirectory(IPCCompiler)
-add_subdirectory(LibURL)
 add_subdirectory(LibWeb)
