--- Meta/Lagom/Tools/CodeGenerators/LibURL/CMakeLists.txt
@@ -1,1 +1,0 @@
-lagom_tool(GeneratePublicSuffixData SOURCES GeneratePublicSuffixData.cpp LIBS LibMain)
